modified:   .eslintrc.js

+++ .eslintrc.js
@@ -14,6 +14,6 @@ module.exports = {
     },
     "rules": {
         "semi": ["error", "always"],
-        "quotes": ["error", "double"]
+        "quotes": ["error", "single"]
     }
 };
